Parcial02_JaimeRodRod/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
dnspython==2.7.0
fastapi==0.115.5
h11==0.14.0
httptools==0.6.4
idna==3.10
motor==3.6.0
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.2
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
